# add CompareTwoPairings