The `Verify` function computes e(P1, Q1) == e(P2, Q2) where e is pairing.
But Pairing(P, Q) = FE(ML(P, Q)) where ML is MillerLoop and FE is FinalExponentiation.
e(P1, Q1) == e(P2, Q2)
<=> e(P1, Q1) e(-P2, Q2) == 1
<=> FE(ML(P1, Q1) ML(P2, Q2)) == 1 then I added `CompareTwoPairings`, which can reduce one FE.
I got the following speed on Core i7-7700 3.6GHz:

* original g1pubs
```
bls/g1pubs% go test --bench .
goos: linux
goarch: amd64
pkg: github.com/phoreproject/bls/g1pubs
BenchmarkBLSAggregateSignature-4                  200000              8234 ns/op
BenchmarkBLSSign-4                                   500           3755491 ns/op
BenchmarkBLSVerify-4                                 100          14766535 ns/op
BenchmarkVerifyWithDomain-4                          100          16862432 ns/op
BenchmarkVerifyAggregateCommonWithDomain-4           100          16845176 ns/op
BenchmarkVerifyAggregateMultipleWithDomain-4           1        1327194394 ns/op
```

* optimized g1pubs
```
pkg: github.com/phoreproject/bls/g1pubs
BenchmarkBLSAggregateSignature-4                  200000              8189 ns/op
BenchmarkBLSSign-4                                   500           3809514 ns/op
BenchmarkBLSVerify-4                                 100          10010070 ns/op
BenchmarkVerifyWithDomain-4                          100          11823728 ns/op
BenchmarkVerifyAggregateCommonWithDomain-4           100          11900917 ns/op
BenchmarkVerifyAggregateMultipleWithDomain-4           1        1349163898 ns/op
```

* original g2pubs
```
bls/g2pubs% go test --bench .
BenchmarkBLSAggregateSignature-4         1000000              2143 ns/op
BenchmarkBLSSign-4                          2000           1073271 ns/op
BenchmarkBLSVerify-4                         100          13116371 ns/op
```

* optimized g2pubs
```
goos: linux
goarch: amd64
pkg: github.com/phoreproject/bls/g2pubs
BenchmarkBLSAggregateSignature-4         1000000              2144 ns/op
BenchmarkBLSSign-4                          2000           1081131 ns/op
BenchmarkBLSVerify-4                         200           8058618 ns/op
```